moved catalog package to the node level
