+++ pkg/controller/nodes/catalog/noop_catalog.go
